refactor: вынести логику получения фильмов в App
